% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure.R
\name{create_figure}
\alias{create_figure}
\title{Create a \code{ggplot2} figure}
\usage{
create_figure(
  dataset = generate_data(size = "strong", threshold = log(0.75)),
  reference = c("none", "lines", "lines + text"),
  y_label = c("index", "change"),
  ci = c("none", "band", "gradient"),
  effect = c("none", "symbol", "colour symbol", "colour ci", "symbol + colour ci")
)
}
\arguments{
\item{dataset}{the output of \code{generate_data()}.}

\item{reference}{Which reference to show.
Must be either \code{"lines + text"} (default), \code{"lines"} or \code{"none"}.}
}
\description{
Create a \code{ggplot2} figure
}
