% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate a time series}
\usage{
generate_data(
  size = c("stable", "moderate", "strong", "potential"),
  threshold = log(0.75)
)
}
\arguments{
\item{size}{The effect size of the strongest effect.
Must be one of \code{"stable"}, \code{"moderate"}, \code{"strong"} or \code{"potential"}.}

\item{threshold}{log threshold value.}
}
\description{
Generate a time series
}
