% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_design.R
\name{generate_design}
\alias{generate_design}
\title{Generate a design with both pairwise comparisons and constant effects}
\usage{
generate_design(design_effect, design_data, max_question = 10)
}
\arguments{
\item{design_effect}{A named list of vectors with design effects.
One character vector with effect levels per design effect.
The output contains paired questions of design effects.}

\item{design_data}{A named list of vectors with design data.
One character vector with data levels per design data.
The output contains a constant value for the data.}

\item{max_question}{The maximum number of questions.
Defaults to 10.}
}
\description{
Generate a design with both pairwise comparisons and constant effects
}
