% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_app_data.R
\name{import_app_data}
\alias{import_app_data}
\title{Import the results from the app}
\usage{
import_app_data(root, target)
}
\arguments{
\item{root}{the data folder of the app}

\item{target}{the folder were to store the aggregated data}
}
\description{
Import the results from the app
}
