% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{scale_data}
\alias{scale_data}
\title{Generate a time series}
\usage{
scale_data(f_down = -0.75, f_up = 0.25, threshold = log(0.75))
}
\arguments{
\item{f_down}{Relative location of the lower confidence limit of the
strongest effect.
Relative to the \code{threshold}.}

\item{f_up}{Relative location of the upper confidence limit of the
strongest effect.
Relative to the \code{threshold}.}

\item{threshold}{log threshold value.}
}
\description{
Generate a time series
}
