% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_figure.R
\name{create_figure}
\alias{create_figure}
\title{Create a \code{ggplot2} figure}
\usage{
create_figure(
  dataset = generate_data(size = "strong", threshold = log(0.75)),
  reference = c("none", "lines", "lines + text"),
  y_label = c("index", "change"),
  ci = c("none", "band", "gradient"),
  effect = c("none", "symbol", "colour symbol", "colour ci", "symbol + colour ci"),
  scale_points = 1,
  highlight = NULL
)
}
\arguments{
\item{dataset}{the output of \code{generate_data()}.}

\item{reference}{Which reference to show.
Must be either \code{"lines + text"} (default), \code{"lines"} or \code{"none"}.}

\item{y_label}{Defines the style of the y axis labels
\code{"index"} generates labels as rescaled values where the reference is set to
100.
\code{"change"} generates labels as the relative difference to the reference.}

\item{ci}{Defines the style of the confidence intervals.
\code{"none"} hides the confidence intervals.
\code{"band"} displays only the 90\% confidence intervals.
\code{"gradient"} displays the 30\%, 60\% and 90\% confidence intervals.}

\item{effect}{Defines the style of the classification.
\code{"none"} hides the classification.
\code{"symbol"} displays point with text symbols.
\code{"colour symbol"} displays point with text symbols with different colours.
\code{"colour ci"} colours the confidence intervals.
\code{"symbol + colour ci"} colours the confidence intervals and displays points
with text symbols`.}

\item{scale_points}{a scaling factor for point symbols.
Defaults to \code{1}.}

\item{highlight}{Which year to highlight with a vertical dashed line.}
}
\description{
Create a \code{ggplot2} figure
}
